Update example in README.md
## Why?
While using the package I followed the example in the README but noticed that was not setting the actual cookie in the browser. After digging I noticed that was because it [needs the necessary cookies to be true](https://github.com/use-cookie-consent/use-cookie-consent-core/blob/feeb78cd823b21cb9f626e84526ffcec10b8ef28/src/useCookieConsent.ts#L24) to set the cookie and even when the default value of the necessary cookies is `true`, the initial value does not persist using `acceptCookies()`. The easiest fix was just setting the necessary cookies to `true` along with the new value.

I also added a basic example of the hook with custom cookie attributes.